+++ test/test_helper.rb
@@ -7,7 +7,7 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 ENV["RAILS_ENV"] = "test"
 
 begin
-  require "pry-byebug"
+  require "debug" unless ENV["CI"]
 rescue LoadError
 end